Introduce test framework for LITMUS^RT.
This is the beginning of the LITMUS^RT testsuite.
The main design goals are flexibility and ease of test writing.

To create a new test, simply write a test case in any C file in the
tests/ subdirectory. The buildsystem will find the test and hook it
up with the testrunner.

Have a look at tests/fdso.c and include/tests.h to get an idea for what
tests look like.

Tests can be executed with the 'runtests' tool. Each testcase is executed
in a separate process in order to ensure that tests do not influence
each other.