refactor: switch back from SCons to make
We originally switched from make to scons because
1) our makefiles were not very good;
2) SCons promised to make maintaining the build system simpler.

Unfortunately, SCons has become more and more difficult to deal with
as we moved to supporting several architecture and cross compilation,
to the extend that we ended up re-creating make functionality in SCons.

So let's switch back to make using a "clean" Makefile.

Thanks a lot to Andrea Bastoni and Chris Kenna for feedback on
previous iterations of these patches.