Port Makefile to new Linux kernel header layout
The userspace versions of unistd.h have moved to include/uapi/asm in
recent kernel versions; we need to adjust the liblitmus Makefile
accordingly.

Further, on x86, the unistd_32.h and unistd_64.h files are now
actually being auto-generated; this also forced adjustments.

This patch is required for liblitmus to compile against Linux 3.10.