Add -r (reservation) and -v (verbose) flags to rtspin
The -r flag allows specifying a "virtual CPU" (in addition to a
physcial CPU to migrate to) that may be interpreted by a scheduler
plugin to refer to a reservation to which the task should be added.

The -v flag causes rtspin to output timing information on stdout and
can be useful for early-stage plugin debugging (to see that something
is happening at all).