Introduce a task configuration based on amount of work
In order to make experiments in virtualized environment it is
introduced a task (rtspin) configuration based on amount of work.

Though this configuration requires some manual operation for choosing
number of cycles burned for 1ms of task execution.
First you need to run 'rtspin -A' baremetal or on a virtualized
environment with pinned CPUs with 100% bandwidth and take a cycles number
for 1ms taks. Optionally verify chosen number with 'rtspin -l -a CYCLES'.
Then use the number for workload runs.

Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>